Updated scripts with metrics and improvements
Metrics and improvements in test and catalog
