+++ README.md
@@ -4,7 +4,9 @@ Repository used to generate the JSON-LD and HTML catalogs for FOOPS!
 
 In order for the script to work correctly, the following paths have been configured:
 path_ttls = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/test/'
+path_ttls_metric = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/metric/'
 path_mustache = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/test/template.html'.
+path_mustache_metrics = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/metrics/template_metrics.html'
 
 If the TTL files or the template change location, these paths will need to be updated.
 
@@ -16,7 +18,8 @@ The script requires the rdflib and pystache libraries for proper operation. Both
 
 The main page created from ttl_catalogue.py goes through all test folders and retrieves information to create a catalog item whenever it finds a ttl file in the folder.
 
-It requires specifying the path where the test folders are located and where the template is hosted, from which the catalog will be created.
+It requires specifying the path where the test and metric folders are located and where the template is hosted, from which the catalog will be created to show the list of tests and metrics availables.
 
 path_ttls = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/test/'
+path_ttls_metric = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/metric/'
 path_mustache = 'https://github.com/oeg-upm/fair_ontologies/tree/main/doc/template_catalog.html'.
